feat: run preloaded huggingface models in parallel
Allow some huggingface models to be loaded at boot time and reused as
needed. When the new "preload" field from the model's config in storage
is true, eagerly load the model before serving any HTTP requests.

The core of this feature is in InferenceManager, where we keep a dict
of all preloaded huggingface models and populate it in the constructor.

In bulk_completions we use the InferenceManager to determine which
InferenceRequests we can serve with a preloaded model and which we
would need to load a new model for. Generations using preloaded models
are scheduled in parallel (along with generations by remote models).

During text generation the InferenceManager checks its preloaded cache
reuses the model if one is present.

Also add --force-preload-hf flag and associated FORCE_PRELOAD_HF_MODELS
environment variable. This can be used to set the "preload" field to
true in the config for any huggingface model at startup.